Move car_model.py from optimization to dynamics, adjusting any references
