+++ dynamics/README.md
@@ -7,6 +7,7 @@ A directory for all the car's dynamics modelling and data
 + `tests` directory
 + `CdACrrCalculator.py`: Extract aero drag and rolling resistance coefficients
 + `parserolldowndata.py`: Method to produce cleaned dataframe from CSV rolldown
++ `car_model.py`: Class to model car object, and using free-body diagram physics to determine energy needed for the car parameters and route
 
 ### `motor_efficiency` module
